+++ Changelog.md
@@ -1,6 +1,6 @@
 # Changelog for the `parameterized-utils` package
 
-## next
+## 2.1.10.0 -- *2025 Mar 20*
 
   * New instances for types from `base`: